+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- ...
+- Custom filters to Uri generator.
 
 ## [0.4.0] - 2022-08-03
 ### Added